Split private and public ReactJSXElementValidator tests
